### 2019-06-24 #58 @whedon
[PRE REVIEW]: M4: Metals, Magnets, and Miscellaneous Materials
**Submitting author:** @albi3ro (<a href="http://orcid.org/0000-0002-2127-8453">Christina Colleen Lee</a>)
**Repository:** <a href="https://github.com/albi3ro/M4" target ="_blank">https://github.com/albi3ro/M4</a>
**Version:** v_jose-submit
