### 2019-06-24 #58 @labarba
👋  @albi3ro —You will need to add the missing fields to the paper metadata.
See error: `Paper YAML header is missing expected fields: affiliations, bibliography `
