### 2019-06-24 #58 @labarba
@katyhuff — The author suggested you as handling editor for this JOSE submission, but I know you're already editing another. Let me know if you're willing to do one more... otherwise, perhaps @moorepants will be willing? (although not quite aligned with area of interest/expertise). Help!
