### 2019-06-25 #58 @katyhuff
@albi3ro Please let me know you are able to update the paper metadata to include the missing fields ( `affiliations, bibliography` as suggested by @labarba above) . At that point, I'll begin assigning reviewers.
