### 2019-07-04 c2f8e83
affiliation and bibliography on paper.md
