### 2019-07-04 #58 @albi3ro
I updated `paper.md`, so it should work now.

I don't have an affiliation at the moment, so I created an affiliation named "Independent Researcher".  But if you can a different protocol for no affiliations, I can change that.
