### 2019-07-11 #58 @labarba
Your editor, @katyhuff, will be looking for reviewers for the submission. You could help if you find someone on [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE that you can recommend!
