### 2019-07-11 #58 @katyhuff
Thanks @labarba -- and @albi3ro I'm sorry for the delay. I apologize, I thought you were still working on the bibliography part of the request. @albi3ro:  Please note that your bibliography still doesn't appear (now, simply because you don't reference anything).

In any case, I believe that I have identified a couple of appropriate reviewers whom I'll invite now. If you have suggestions, of course, they are more than welcome.
