### 2019-07-11 #58 @katyhuff
@trallard : Tania Allard, are you able to and interested in taking on this review for JOSE (the Journal of Open Source Education) ? This submission is in need of expertise in materials as well as familiarity with the systems and languages involved (julia, jupyter). I think you would be an excellent reviewer. Please see the [JOSE reviewer guidelines](http://jose.theoj.org/about#reviewer_guidelines) for details about the JOSE review process.

* Title: M4: Metals, Magnets, and Miscellaneous Materials
* Summary: ... the wonders of materials all around you, and some of the revolutions that are shocking the field. Posts are converted jupyter notebooks with Julia code, and are meant for interaction.
* Article Proof:  https://github.com/openjournals/jose-papers/blob/jose.00058/jose.00058/10.21105.jose.00058.pdf
* Submitting author: @albi3ro (Christina Colleen Lee)
* Repository: https://github.com/albi3ro/M4
* Website: http://albi3ro.github.io/M4/
* Version: v_jose-submit
