### 2019-07-11 #58 @trallard
Hey @katyhuff I would love to take this on.

The problem is I have another review that I want to complete before going on leave on the 27th. I can try and make some advances on this (though) I cannot promise anything. Or move it forward after the 14th of August.

Would this work for y'all?
