### 2019-07-12 #58 @labarba
I wonder if @IanHawke could help with this review ... Ian—this submission to JOSE consists of Jupyter notebooks using Julia, addressing physics of materials. Could you have a look and consider if you could contribute a review?
