### 2019-07-12 #58 @labarba
Another person who might be able to help with this JOSE review is @dga_smith (more chemistry than physics background, I think, but materials science sits at some intersection of those fields).

@dga_smith — Could you have a look at this JOSE submission and consider if you could contribute a review? Or perhaps you can think of someone else who could?
[ Here's the article proof :page_facing_up: :point_left: ](https://github.com/openjournals/jose-papers/blob/jose.00058/jose.00058/10.21105.jose.00058.pdf)
