### 2019-07-12 #58 @IanHawke
Technically I think I can do this, yes (with some brushing up). As far as timing goes I will only be able to start toward the end of next week, and probably only have time to cover the straightforward part of the checklist before having to stop for a week. I would hope to be able to comment on all the material by the end of the month. Sorry I can't promise anything faster.
