### 2019-07-12 #58 @lheagy
Thanks for thinking of me @katyhuff; this looks like an interesting set of resources! I am a bit conference over-loaded for the next few weeks, so realistically it would likely be mid-august before I could take a look... Please feel free to keep me on the list as a back-up, if you are still in need of someone in a few weeks time
