### 2019-07-22 d2f91a5
Correct JuliaBox link, add image.
