+++ README.md
@@ -22,14 +22,18 @@ I vary the level from undergraduate physics major to graduate student in the fie
 
 The blog will be written in the form of Julia (http://julialang.org/) Jupyter notebooks. This language has a syntax similar to MATLAB or Python, and is easily human readable.  I encourage active participation by playing with the code.
 
-You can either install Julia and Jupyter notebook on your own machine (https://julialang.org/downloads/), or use JuliaBox (https://www.juliabox.org/) to view and interact with the files. I've recently updated the notebooks to version v1.0.1, though any of the v1.0.x should work.
+You can either install Julia and Jupyter notebook on your own machine (https://julialang.org/downloads/), or use JuliaBox (https://www.juliabox.com/) to view and interact with the files. I've recently updated the notebooks to version v1.0.1, though any of the v1.0.x should work.
 
-To use JuliaBox (the easiest option), log in with a Google account, and then click on the sync tab.  Under the Github directory, input the clone https://github.com/albi3ro/M4.git , use branch master, and name the folder as you see fit.
+To use JuliaBox (the easiest option), log in with a Google account, and then click on the sync (Git) tab.  
+
+![The JuliaBox launch page](images/juliabox.png)
+
+Under the Github directory, input the clone https://github.com/albi3ro/M4.git , use branch master, and name the folder as you see fit.
 
 Check out http://julialang.org/ for documentation on this new and exciting language, and learn how to install it on your own machine.
 
 ### Introductory note on packages
-Many times I will use extra packages developed for Julia, such as for plotting functionality. If you are using the package for the first time on your machine, you will need to evaluate 
+Many times I will use extra packages developed for Julia, such as for plotting functionality. If you are using the package for the first time on your machine, you will need to evaluate
 
 ```julia
     using Pkg
@@ -46,9 +50,9 @@ You can also install a package at the REPL (Read Evaluate Print Loop) before ent
 ### Using my packages
 I have written packages of my own for use in certain sections.  To add these to the path, download them to the folder you are using, and in your command line, type `push!(LOAD_PATH,"/path/to/my/code")`, where `/path/to/my/code` is the location of the package.  Then, you can type `using Lattices` in just the same way.  
 
-### Contributing 
+### Contributing
 
-If you have expierence in a topic you would wish to see covered, please email me (chrissie.c.l@gmail.com).  I have multiple topics I would love to cover, but just don't have the expertise or time to learn everything myself from scratch. 
+If you have expierence in a topic you would wish to see covered, please email me (chrissie.c.l@gmail.com).  I have multiple topics I would love to cover, but just don't have the expertise or time to learn everything myself from scratch.
 
 If you spot any grammar issues, feel free to either give a pull request or start an issue.  
 
