### 2019-07-22 albi3ro/M4#9 @IanHawke
Earlier images in root-finding notebook
Linked to https://github.com/openjournals/jose-reviews/issues/58 .

This relates to https://github.com/albi3ro/M4/blob/master/Numerics/Roots_1D.ipynb .

The top of the notebook emphasises (correctly IMO) the importance of graphing the function. However, there's no pictorial examples until quite late: this makes the *Multiplicity* section not so easy to understand. In addition, there's no comment on what happens with multiple roots.

Suggest that the functions are introduced and plotted earlier, and that functions with multiple roots are also discussed. On the range you're using I'd suggest `cos(2*pi*x)-0.4*x` which has around 10 roots between -5 and 5.
