### 2019-07-22 #58 @william-pfalzgraff
Hi @katyhuff. I'm currently on vacation, but I'd be able to start my review next week (the week of July 29).  Would that work?
