### 2019-07-22 #61 @whedon
[REVIEW]: M4: Metals, Magnets, and Miscellaneous Materials
**Submitting author:** @albi3ro (<a href="http://orcid.org/0000-0002-2127-8453">Christina Colleen Lee</a>)
**Repository:** <a href="https://github.com/albi3ro/M4" target ="_blank">https://github.com/albi3ro/M4</a>
**Version:** v_jose-submit

- [ ] **Version:** Does the release version given match the repository release (v_jose-submit)?

- [x] **Version:** Does the release version given match the repository release (v_jose-submit)?
