### 2019-07-22 #58 @katyhuff
@albi3ro Thank you for your patience. I've now found reviewers, and we'll be able to start the review discussion over in #61. Please keep an eye on that issue for comments from reviewers in the next couple of weeks.

@william-pfalzgraff @IanHawke Thank you for volunteering! Let's move over to #61 where we'll conduct the review!
