### 2019-08-10 eb72640
update package section of README
