### 2019-08-15 #61 @IanHawke
As an update on where this stands with me:

* Two PRs linked in the pre-review thread (https://github.com/albi3ro/M4/pull/7 , https://github.com/albi3ro/M4/pull/8) have been merged. The one about JuliaBox needs revisiting by me, as whilst the PR fixed the README, this change needs propagating to the front page of the website as well.
* One issue linked in the pre-review thread (https://github.com/albi3ro/M4/issues/9) has not yet been addressed yet, but there's lots more content stuff to look at.
* I've gone through the first two sections (General, Prerequisites Required) and have a number of minor comments but few significant problems. Issues and PRs to come.
* I think the paper is missing some of the key points above, particularly the **Description** (making the sequence clear will be an issue/PR soon) and the story (as some blog pages suggest that sections were used in classroom contexts, but others seem aimed at remote audiences).
* I need to dig further into how the webpages are generated, as there are a few where the pages differ from the notebooks in substantive ways.

Hope to get through the rest of the material next week.
