### 2019-08-15 albi3ro/M4#10 @IanHawke
Non-running notebooks
Linked to https://github.com/openjournals/jose-reviews/issues/61 .

Having checked the "General" and "Prereqs" section, the following notebooks didn't run for me currently (Julia v 1.1, Mac):

- [ ] https://github.com/albi3ro/M4/blob/master/Prerequisites/Time%20Evolution.ipynb : Needs to import `FFTW`, and a number of places need dot operators adding (eg, cell 4 needs `abs(x)` changing to `abs.(x)`, I think).
- [ ] https://github.com/albi3ro/M4/blob/master/Prerequisites/Imaginary%20Time%20Evolution.ipynb : Didn't run first time, but did later? May be that `FFTW` isn't in the current environment file or I just messed up.
- [ ] https://github.com/albi3ro/M4/blob/master/Prerequisites/Atomic%20Orbitals.ipynb : Error in cell 9 (a plotting issue I don't understand yet).
- [ ] https://github.com/albi3ro/M4/blob/master/Prerequisites/Atomic%20Orbitals%202.ipynb : Needs to `import Pkg` at the top, Add dots in `CarttoSph`, and doesn't seem to like `e` (?: I defined it as `exp(1)` to make it run). As this relies on `GLVis` in the end, which doesn't install for me, I'm only trying to get the functions to run.
