### 2019-08-15 albi3ro/M4#11 @IanHawke
Broken links on static pages
Linked to https://github.com/openjournals/jose-reviews/issues/61 .

Note that this is, in some ways, a duplicate of https://github.com/albi3ro/M4/issues/3 .

Some of the internal links on the static pages are broken. For example, the [Phase Transitions](http://albi3ro.github.io/M4/prerequisites%20required/Phase-Transitions.html) page link to the [MC Ferromagnet](http://albi3ro.github.io/M4/prerequisites%20required/Monte-Carlo-Ferromagnet.html) page points to the wrong place (the directory is `prerequisites%20required` instead of `prerequisites`).

From what I can tell it's only links to Prereqs that cause problems now, but if it's the space that's causing the issue then links to General ("General audience") may have the same problem in the future.
