### 2019-08-15 albi3ro/M4#12 @IanHawke
Make notebook order clear
Linked to https://github.com/openjournals/jose-reviews/issues/61 .

At the moment there are effectively two reading orders; the one from the [front page of the static blog](http://albi3ro.github.io/M4/), and the one in the [table of contents](https://github.com/albi3ro/M4/blob/master/TableofContents.md).

The front page of the blog can't be the right order, as "Part 2" often comes before "Part 1" (this is particularly clear in the Prereqs section). The table of contents is a lot clearer on this, but isn't referred to anywhere.

I would suggest  putting an `index.ipynb` at the top level mirroring the content of the table of contents and see if there's a way of generating the static front page from that. This is, in part, because (combined with the environment) it may be possible to run the site through https://mybinder.org as an alternative to JuliaBox.

Either way, the order needs to be clearer and linked from the README as well as fixed on the blog front page.
