+++ _config.yml
@@ -3,7 +3,7 @@ markdown: kramdown
 highlighter: rouge
 allurl: albi3ro.github.io/M4
 baseurl: /M4
-permalink: none
+permalink: /:title:output_ext
 gems:
   - jekyll-feed
 
