+++ _posts/2017-07-11-Spin-Chain-pt2.md
@@ -10,7 +10,7 @@ image: "https://albi3ro.github.io/M4/Images/SpinChainED/spinchain.png"
 description: Construct the matrix corresponding to the 1D Heisenburg Spin Chain and calculate its eigenvalues and eigenvectors.
 ---
 
-Check out [Part 1]({{base.url}}/M4/graduate/Spin-Chain-Prerequisites.html) for the theoretical background.  Today is all programming.
+Check out [Part 1]({{base.url}}/M4/Spin-Chain-Prerequisites.html) for the theoretical background.  Today is all programming.
 
 Remember to keep a reasonable number of spins, $n$.  The number of states goes as $2^n$, and the size of the Hamiltonian will go as $2^n \times 2^n$.  A 10-spin chain will have over a million entries, before taking into account any memory reducing tricks.
 
