+++ _posts/2015-01-04-Atomic-Orbitals.md
@@ -14,7 +14,7 @@ Electrons around a nucleus.  Do they look like little well behaved planets orbit
 
 NOPE!
 
-We get spread out blobs in special little patterns called orbitals.  Here, we will look at their shapes and properties a bit.  Today we will look at graphs in 1D and 2D, but the next post, [Atomic Orbitals Pt. 2]({{base.url}}/M4/prerequisites/Atomic-Orbitals2.html), uses a fancy, but slightly unstable plotting package, GLVisualize to generate some 3D plots.
+We get spread out blobs in special little patterns called orbitals.  Here, we will look at their shapes and properties a bit.  Today we will look at graphs in 1D and 2D, but the next post, [Atomic Orbitals Pt. 2]({{base.url}}/M4/Atomic-Orbitals2.html), uses a fancy, but slightly unstable plotting package, GLVisualize to generate some 3D plots.
 
 The Hamiltonian for our problem is:
 
@@ -272,4 +272,4 @@ colorbar()
 {% include image.html img="M4/Images/Orbitals/angular2di.png" title="3p in 2d" caption="Slice of a 3p orbital in the x and z plane."%}
 {% include image.html img="M4/Images/Orbitals/angular2d2i.png" title="3dz2 in 2d" caption="Slice of a 3dz2 orbital in the x and z plane."%}
 
-Don't forget to check out [Atomic Orbitals Pt. 2]({{base.url}}/M4/prerequisites/Atomic-Orbitals2.html)!
+Don't forget to check out [Atomic Orbitals Pt. 2]({{base.url}}/M4/Atomic-Orbitals2.html)!
