+++ jupyterjulia.md
@@ -10,7 +10,7 @@ The blog will be written in the form of julia (<a href="http://julialang.org/">h
 
 Check out <a href="http://julialang.org/">http://julialang.org/</a> for documentation on this new and exciting language, and learn how to install it on your own machine.  "Docs" on the right hand side provides an easy link to the documentation.  
 
-You can either install Julia and jupyter notebook on your own machine, or use juliabox (<a href="https://www.juliabox.org/">https://www.juliabox.org/</a>) to view and interact with the files.
+You can either install Julia and jupyter notebook on your own machine, or use juliabox (<a href="https://www.juliabox.com/">https://www.juliabox.com/</a>) to view and interact with the files.
 
 To use juliabox (the easiest option), log in with a google account, and then click on the sync tab.  Under the Github directory, input the clone <a href="https://github.com/albi3ro/M4.git">https://github.com/albi3ro/M4.git</a> , use branch master, and name the folder as you see fit.
 
