+++ _posts/2015-01-08-MCMC.md
@@ -7,7 +7,7 @@ tags: [Monte Carlo]
 author: Christina C. Lee
 ---
 
-<b>Prerequisites:<b/>  [Monte Carlo Calculation of pi]({{base.url}}/M4/numerics/Monte-Carlo.html)
+<b>Prerequisites:<b/>  [Monte Carlo Calculation of pi]({{base.url}}/M4/Monte-Carlo.html)
 
 <b>Author:</b> Christina Lee
 
