### 2019-08-19 deced76
renaming, notebook toc, updating simplex, removing deprecations
