### 2019-08-23 albi3ro/M4#10 @albi3ro
I did find some issues in "Time Evolution" that I fixed.
I didn't find any issues with "Imaginary Time Evolution", but I had gone through Time Evolution and had added FFTW (so now FFTW is in packages list).
Those two should work now. Let me know if they don't.

Going over "Atomic Orbitals", I realize it does generally have problems that I am now trying to fix.  I originally used the package `PyPlot.jl` to plot with, then switched to `Plots.jl`. I changed the syntax, but I didn't change what I was doing or how I was using the package.  I'm working on that now.

"Atomic Orbitals 2" used a confusing and now deprecated package `GLVisualize.jl`.  I've just fully deleted it now.
