### 2019-08-23 a132f94
One notebook using LaTeX labels
