### 2019-08-23 albi3ro/M4#10 @IanHawke
Those all work, thanks.

Here come the next lot:

- [ ] https://github.com/albi3ro/M4/blob/master/Numerics_Prog/HDF5.ipynb : `dump()` returns `nothing` which causes `println` to break. If all `println(dump(...))` lines are replaced with `dump(...)` it works for me. Note that it may make sense to split the cells so the `close(fid)` line is in a separate cell: this means the later cells can be run even if an earlier cell breaks. However, I'm not sure this is a good suggestion, as keeping it in the same cell implicitly emphasizes that `close` should be done ASAP.
- [ ] https://github.com/albi3ro/M4/blob/master/Graduate/Chern-Number.ipynb : `linspace` is now `range`, needs `using LinearAlgebra` to get `norm`, and I think the `Array` definition syntax needs fixing in places.

I think this is it, in terms of getting the notebooks running.
