### 2019-09-13 157a78c
rewriting orbitals, fixing 1d roots, switching to plotlyjs, others
getting personal lattices module working, general editing, graphics
