+++ Project.toml
@@ -10,7 +10,7 @@ GSL = "92c85e6c-cbff-5e0c-80f7-495c94daaecd"
 Gadfly = "c91e804a-d5a3-530f-b6f0-dfbca275c004"
 GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
 HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
-Lattices = "c21dee59-005f-55b6-acfd-94526b082a96"
+LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 ORCA = "47be7bcc-f1a6-5447-8b36-7eeeff7534fd"
