+++ _posts/2019-03-21-Roots_1D.md
@@ -40,13 +40,62 @@ Mostly, figure out as much behavior beforehand as you can.
 * Do we have a derivative? 
 * Could we have an even multiplicity situation?
 
+```julia
+using ForwardDiff
+using Statistics
+using SpecialFunctions
+using Plots
+plotly()
+```
+
 ### Multiplicity
 
-The function $x^2-1$ intersects the x-axis twice.  As we move it up, at the function $x^2$ only intersects once, but we consider it a <i>zero of multiplicity two</i>. We count the zero of $x^2$ as an ''axis-crossing going down'' and an ''axis-crossing going up'' both at the same time.  
+The function $x^2-1$ intersects the x-axis twice.
 
-If we have multiplicity three like $x^3$ at zero, the function goes up, then down, then back up again.  Overall, $x^3$ heads straight across the x-axis.  
+```julia
+    plot(x->x^2-1,label="x^2-1")
+    hline!([0],label="")
+    plot!(xlims=(-2,2),ylims=(-2,5))
+```
 
-Again at another even multiplicity, $x^4$ with multiplicity 4, the function goes down, up, down, then up.  Overall, $x^4$ hits the axis but never crosses it.  
+<iframe src="/M4/Images/Roots/multiplicity1.html"  style="border:none; background: #ffffff"
+width="700px" height="450px"></iframe>
+
+As we move the graph up, at the function $x^2$ only intersects once, but we consider it a <i>zero of multiplicity two</i>, or "two zeros at the same point". We count the zero of $x^2$ as an ''axis-crossing going down'' and an ''axis-crossing going up'' both at the same time.  
+
+```julia 
+    plot(x->x^2,label="x^2")
+    hline!([0],label="")
+    plot!(xlims=(-2,2),ylims=(-2,5))
+```
+
+<iframe src="/M4/Images/Roots/multiplicity2.html"  style="border:none; background: #ffffff"
+width="700px" height="450px"></iframe>
+
+
+If we have multiplicity three like $x^3$ at zero, the function goes up, then down, then back up again.  Overall, $x^3$ heads straight across the x-axis.  We see these three zeros seperately if we perturb the function with another term, like $-2x$.
+
+```julia
+    plot(x->x^3,label="x^3")
+    plot!(x->x^3-2x,label="x^3-2x")
+    hline!([0],label="")
+    plot!(xlims=(-2,2),ylims=(-2,5))
+```
+
+<iframe src="/M4/Images/Roots/multiplicity3.html"  style="border:none; background: #ffffff"
+width="700px" height="450px"></iframe>
+
+Again at another even multiplicity, $x^4$ with multiplicity 4, the function goes down, up, down, then up.  Overall, $x^4$ hits the axis but never crosses it. 
+
+```julia
+    plot(x->x^4,label="x^4")
+    plot!(x->x^4-3x^2+1,label="x^4-3x^2+1")
+    hline!([0],label="")
+    plot!(xlims=(-2,2),ylims=(-2,5))
+```
+
+<iframe src="/M4/Images/Roots/multiplicity4.html"  style="border:none; background: #ffffff"
+width="700px" height="450px"></iframe>
 
 When we have zeros of even multiplicity, we will <b>hit the axis</b> $f(x)=0$, but <b>never change signs</b>.  
 
@@ -62,15 +111,6 @@ In this post, we cover:
 * Regula-Falsi
 
 
[…]
