### 2019-09-13 4b830bd
removing simplex stuff
yep.... helpless post and it's  time to prune.
