+++ _posts/2017-07-06-Homology-pt1.md
@@ -1,432 +1,0 @@
----
-title: Homology Part 1
-layout: post
-comments: True
-category: Graduate
-tags: Mathematics
-author: Christina C. Lee
-description: What is a simplicial complex and what do they look like?
-image: "https://albi3ro.github.io/M4/Images/simplex/simplex.png"
----
-
-I've been learning the pure mathematics of topology to understand the physics of topological materials better.   Predictably, my sources have a lot of theorems, the examples tend to be esoteric and hard to conceptualize, and I have little idea how it all relates to my pretty crystals.  There also aren't that many examples.  I almost miss the days back in high school when they forced us to do the same problem 1,000 times over.
-
-So to better understand one of the important topics, simplices, and simplicial complexes, I've decided to take the pure math and put it to programming the best that I can.  I don't know if this is the best way of doing things, or if it's useful at all, but I think it helps me understand them.  Hopefully, it will help you too.
-
-<b>Topology</b> is the study of deforming things.  If we push and pull at something, without tearing or gluing, what properties remain the same?  A coffee cup is the same as a donut because they both have one hole.
-
-We are also the same as a donut because of our digestive track... if you ignore all the open cavities we have inside.  We can at least deform away our lungs.  I'm not an anatomy expert.
-
-Many objects of interest are <b>topologies</b>, which are just collections of sets that obey certain constraints.  Every manifold you've ever worked with has been a topology, but it probably was just trivial and didn't need further study.  To study more complex topologies, we can find an equivalent <b>simplicial complex</b>.  We can then compute a variety of things about the simplicial complex that tells us about the properties of the topology.
-
-Simplicial complexes are made out of <b>simplices</b>.
-
-A <b>simplex</b> is an $n$-dimensional generalization of a triangle.  A 0-dim simplex is a point; a 1-dim simplex is a line; 3-dim a tetrahedron; so on and so forth.
-
-{% include image.html img="M4/Images/simplex/simplex.svg" title="Simplices" caption="Left: Different dimension simplices.  Right: A simplicial complex." %}
-
-The rules for the simplices in a legitimate simplicial complex are
-
-        1. Every point must belong to at least one simplex
-        2. A point can belong to only a finite number of simplices
-        3. Two different simplices either have no points in common, or
-            1. on is a face, or edge, or vertex of the other
-            2. the set of points in common is the whole of a shared face, or edge, or vertex.
-
-[1]
-
-A simplicial complex of dimension $d$ will contain a finite number of $d$-dimensional simplices, the faces of those simplices,  the faces of the faces, and so on until you get to all the points contained in the original simplices.
-
-I will denote each point by a string of length 1, like `"a"`.
-
[…]
