### 2019-09-13 0122ca2
removing gadfly and orbitals2 from site
both deprecated, so why should I keep them around?
