+++ _posts/2015-01-12-Gadfly.md
@@ -1,412 +1,0 @@
----
-title: Gadfly
-layout: post
-comments: True
-category: Programming
-author: Christina C. Lee
-image: "https://albi3ro.github.io/M4/Images/Gadfly/p17.png"
-tags: [Plot]
-description: Moving from Python to Julia, you, like me, might jump straight to PyPlot to fill your plotting needs. But don't jump too fast and overlook the wonders of Gadfly.
----
-
-# I'm Moving to Gadfly
-
-### Christina Lee
-
-So far in this blog, all my plots have either used PyPlot, or the experimental GLVisualize. BUT, NO LONGER!
-
-While writing up my research notes, I've been trying to improve my PyPlot graphs. I was hoping for them to look nicer; Julia's PyPlot was saying no.  We had a falling out over background colors and resolution. So, I have decided to move on to a  plotting partner more willing to cope with my demands.
-
-When I first started learning Julia, Gadfly seemed too unfamiliar, and control too tricky. I didn't give it much further thought.  After some digging, Gadfly has changed my mind, and I hope today I can change yours too.  Right now, like Julia, Gadfly doesn't have the maturity or documentation of PyPlot, but once it does, you better watch out.
-
-## Let's plot some graphs!
-
-
-
-
-```julia
-# Time to load our Packages
-#Pkg.update()
-#Pkg.add("Gadfly")
-using Gadfly
-
-#Pkg.add("Compose")
-using Compose
-```
-
-
-```julia
-# Some test data
-x=collect(1:.4:10);
-y1=sin(x);
-y2=cos(x);
-```
-
-
-```julia
-# A simple plot
-plot(x=x,y=y1)
-```
-
-<iframe src="/M4/Images/Gadfly/p1.js.svg"  style="border:none; background: #ffffff"
-width="600px" height="600px"></iframe>
-
-
-
-Unlike PyPlot, we can give `plot` functions as well as points.  After receiving a function, a max value, and a min value, Gadfly figures everything else out for itself.
-
-For the function, we can pass in either an inbuilt function, a user defined function, or an anonymous function.
-
-The brackets [ ] allow us to group multiple functions together in the plot.
-
-
-```julia
-# plotting function
-
-function testfunction(x::Real)
-    return 1-x^2
-end
-
-plot([cos,testfunction,x->x^2],0,1)
-```
-<iframe src="/M4/Images/Gadfly/p2.js.svg"  style="border:none; background: #ffffff"
-width="600px" height="600px"></iframe>
-
-
-
-But that's just one set of points...
-How do we plot more than one set of data?  That's where layers come in.  Instead of using `plot`, we use `layer`, and set it as a variable.  We can then `plot` those layers.
-
-
-```julia
-p1=layer(x=x,y=y1,Geom.point)
-p2=layer(x=x,y=y2,Geom.point)
-plot(p1,p2)
-```
-
-<iframe src="/M4/Images/Gadfly/p3.js.svg"  style="border:none; background: #ffffff"
-width="600px" height="600px"></iframe>
-
-
-
-
-
-
-
-## Different style of plotting: Geom
-Now, what if we want to plot something other than lines?
-Gadfly allows a wide variety of other ways to plot our data.  We control this with the `Geom` (Geometry) object.
-
-
-```julia
-plot(x=x,y=y1,Geom.point,Geom.line)
-```
-
[…]
