+++ _posts/2015-01-05-Atomic-Orbitals2.md
@@ -1,186 +1,0 @@
----
-title: Atomic Orbitals Pt. 2
-layout: post
-comments: True
-category: Prerequisites Required
-tags: [Quantum]
-author: Christina C. Lee
-description: We try out the GLVisualize.jl package to view the hydrogen orbitals computed in the last post.
-image: "https://albi3ro.github.io/M4/Images/Orbitals2/3p_surface.png"
----
-<b>Prerequiresites:</b> Quantum Mechanics course
-
-If you haven't read it already, check out [Atomic Orbitals Pt. 1]({{base.url}}/M4/Atomic-Orbitals.html).  Today, we try and make some prettier pictures. GLVisualize is quite a beautiful package, but not entirely the easiest to use at this point with some not so consistent documentation.
-
-To add this package:
-
-
-{% highlight julia %}
-Pkg.add("GLVisualize")
-{% endhighlight %}
-
-and test with:
-
-
-{% highlight julia %}
-Pkg.test("GLVisualize")
-{% endhighlight %}
-
-But, other steps may be necessary to get the package working.  On a Mac, I was required to install the Homebrew.jl package.
-
-
-{% highlight julia %}
-#Pkg.update();
-#Pkg.add("GSL");
-using GSL;
-using GLVisualize;
-{% endhighlight %}
-
-
-{% highlight julia %}
-a0=1; #for convenience, or 5.2917721092(17)×10−11 m
-
-# The unitless radial coordinate
-ρ(r,n)=2r/(n*a0);
-
-#The θ dependence
-function Pmlh(m::Int,l::Int,θ::Real)
-    return (-1.0)^m *sf_legendre_Plm(l,m,cos(θ));
-end
-
-#The θ and ϕ dependence
-function Yml(m::Int,l::Int,θ::Real,ϕ::Real)
-    return  (-1.0)^m*sf_legendre_Plm(l,abs(m),cos(θ))*e^(im*m*ϕ)
-end
-
-#The Radial dependence
-function R(n::Int,l::Int,ρ::Real)
-    if isapprox(ρ,0)
-        ρ=.001
-    end
-     return sf_laguerre_n(n-l-1,2*l+1,ρ)*e^(-ρ/2)*ρ^l
-end
-
-#A normalization: This is dependent on the choice of polynomial representation
-function norm(n::Int,l::Int)
-    return sqrt((2/n)^3 * factorial(n-l-1)/(2n*factorial(n+l)))
-end
-
-#Generates an Orbital Funtion of (r,θ,ϕ) for a specificied n,l,m.
-function Orbital(n::Int,l::Int,m::Int)
-    if (l>n)    # we make sure l and m are within proper bounds
-        throw(DomainError())
-    end
-    if abs(m)>l
-       throw(DomainError())
-    end
-    psi(ρ,θ,ϕ)=norm(n, l)*R(n,l,ρ)*Yml(m,l,θ,ϕ);
-    return psi
-end
-
-#We will calculate is spherical coordinates, but plot in cartesian, so we need this array conversion
-function SphtoCart(r::Array,θ::Array,ϕ::Array)
-    x=r.*sin(θ).*cos(ϕ);
-    y=r.*sin(θ).*sin(ϕ);
-    z=r.*cos(θ);
-    return x,y,z;
-end
-
-function CarttoSph(x::Array,y::Array,z::Array)
-    r=sqrt(x.^2+y.^2+z.^2);
-    θ=acos(z./r);
-    ϕ=atan(y./x);
-    return r,θ,ϕ;
-end
-
-"Defined Helper Functions"
-{% endhighlight %}
-
-
-
[…]
