### 2019-09-14 albi3ro/M4#10 @albi3ro
Just fixed the HDF problem.

Seems `dump()` no longer is a function, at least one I could find in the [documentation](https://github.com/JuliaIO/HDF5.jl/blob/master/doc/hdf5.md) .

Try the most recent version of Chern Number.  I think I have `range` and the correct `Array` everywhere now, and am using `abs2.()` instead of `norm()`.
