### 2019-09-14 albi3ro/M4#12 @albi3ro
So I have grouped the posts into series that should be read into a particular order.  For the notebooks, I have a `TableofContents.ipynb` file that lists these out, as well as the links within each particular notebook that exists within a series.

I have now also added this functionality to the website as well.
