### 2019-09-14 albi3ro/M4#9 @albi3ro
I've added pictures and added a multiple roots example.

I'm still contemplating a rewrite of the section on "multiplicity" to remove the word "multiplicity" and just discuss "does it change signs?" and "does the derivative vanish?", but haven't quite put the pieces together yet.
