### 2019-10-20 #61 @labarba
@katyhuff — What's the status with this review? I see the reviewers have checked off a few items, but many remain unchecked. The last reviewer comment was more than a month ago. Could you check with the author if they are working on revising?
