### 2019-10-20 #61 @katyhuff
@albi3ro -- Since a number of issues filed by reviewer @IanHawke remain open in your repository, I have been assuming you're actively trying to make progress on them. Do you expect to be able to respond to these issues soon?
Thank you,
Katy
