+++ README.md
@@ -15,18 +15,15 @@ I vary the level from undergraduate physics major to graduate student in the fie
     the orbitals of a hydrogen-like atom, I will assume you analytically solved it in a Quantum Mechanics course.
     Even if you haven't taken  the classes, feel free to still try.
 * <b>Graduate:</b> Specialty technique that assumes you have a solid grounding in the basics.  I still try to not take any knowledge for granted, but not everyone will be trying to learn these things in the first place.
-* <b>Numerics:</b>Without a specific physics goal.  Sometimes I may just want to talk about something programming or numerical method related.
-* <b>Programming:</b> Something related to the language itself.
+* <b>Numerics and Programming:</b>Without a specific physics goal.  Sometimes I may just want to talk about something programming or numerical method related.
 
 ## Julia
 
 The blog will be written in the form of Julia (http://julialang.org/) Jupyter notebooks. This language has a syntax similar to MATLAB or Python, and is easily human readable.  I encourage active participation by playing with the code.
 
-You can either install Julia and Jupyter notebook on your own machine (https://julialang.org/downloads/), or use JuliaBox (https://www.juliabox.com/) to view and interact with the files. I've recently updated the notebooks to version v1.0.1, though any of the v1.0.x should work.
+You can either install Julia and Jupyter notebook on your own machine (https://julialang.org/downloads/). Any Julia version over v1 should work.  v1.0.5 is the current LTS, or Long Term Support, version, while v1.3.1 is the most recent stable version as of my updating this page.
 
-To use JuliaBox (the easiest option), log in with a Google account, and then click on the sync (Git) tab.  
-
-![The JuliaBox launch page](images/juliabox.png)
+Previously, I recommended Juliabox, but due to the increasing popularity, that service is now subscription based.
 
 Under the Github directory, input the clone https://github.com/albi3ro/M4.git , use branch master, and name the folder as you see fit. If this isn't working automatically, you may need to link an SSH key between your JuliaBox account and Github account.  Basically, copy and paste the big, ugly number from the JuliaBox settings into your [Github Settings](https://github.com/settings/keys).
 
@@ -57,8 +54,6 @@ If you don't want to use an REPL Package `]` mode, you can also add packages ind
 
 See more at the [Pkg documentation](https://julialang.github.io/Pkg.jl/v1/environments/#Using-someone-else's-project-1).
 
-If you are using JuliaBox, you will need to evaluate the those lines each time.
-
 [Plots.jl](https://github.com/JuliaPlots/Plots.jl), the plotting utility I prefer, calls to other plotting packges, which also need to be installed.  I prefer [GR.jl](https://github.com/jheinen/GR.jl) or [PlotlyJS.jl](https://github.com/sglyon/PlotlyJS.jl), though PlotlyJS doesn't seem to be working anymore.
 
 ### Using my packages
@@ -75,13 +70,3 @@ For technical mistakes, please submit a pull request.
 The website [albi3ro.github.io/M4](albi3ro.github.io/M4) has Discuss commenting functionality attached to each page where you can comment.  
 
 You can also contact me (chrissie.c.l@gmail.com) with feedback or questions.
-
-## For further reference
-Interested in learning more?
-
-* Principles of Condensed Matter Physics - Chaikin & Lubensky
-* Condensed Matter in a Nutshell - Mahan
-* Inna Vishik on Quora https://www.quora.com/profile/Inna-Vishik
-* https://thiscondensedlife.wordpress.com/
-
-Know any good ones? Let me know!
