### 2021-02-19 #61 @labarba
Hi @albi3ro 👋  — do you think that you would like to finish this stalled submission for JoSE? Or have you moved on to other activities and are no longer interested? We've all been buried under a rubble of piled up work, and overall world chaos, so just update us on your plans with this, when you can.
