### 2021-05-23 #61 @labarba
Hi @albi3ro — I'm sensitive to the fact that the past 14 months have brought many challenges (and we have been very slow with our editorial duties). Could you let us know if you plan to come back to this JOSE submission, and are interested in seeing through publication? If you are not, then we can withdraw it. We cal also pause it, or set a reminder for a later date.
